requirements.txt
click>=7.0,<8
pip-check-reqs>=2.0.1,<3
pip-licenses>=1.7.1,<2
pylint>=1.8.4,<2
pyyaml~=3.12
pytest>=3.4.0,<4
pytest-cov>=2.5.1,<3
pytest-pythonpath>=0.7.2,<1
setuptools>=38.4.0
Sphinx>=2.2.0
sphinx-rtd-theme>=0.4.3,<1
tox>=3.0.0,<4
twine>=1.11.0,<2
nbformat>=5.0.4
panel==0.8.3
plotly==4.0.0
tabulate==0.8.7
pandas==0.24.2

# ce_api requirements
certifi>=14.05.14
six>=1.10
python_dateutil>=2.5.3
urllib3>=1.15.1

absl-py==0.8.1
avro_python3==1.8.1
apache_beam[gcp]==2.17
pyarrow==0.15.1
tensorflow==2.1.0
tensorflow_serving_api==2.1.0
tensorflow_model_analysis==0.21.4
tfx-bsl==0.21.4